refactor(identity): retire Batch 3 auth and user proxy files and align SSOT imports
