+++ backend/api/src/controllers/invoice/invoiceMutationController.ts
@@ -1,7 +1,7 @@
 import logger from '@esparex/core/utils/logger';
 import { Request, Response } from 'express';
 import { createInvoiceRecord } from '@esparex/core/services/InvoiceService';
-import { findUserByEmail } from '@esparex/core/services/UserService';
+import { findUserByEmail } from '@esparex/core/domains/identity/application/users/UserService';
 import { sendErrorResponse } from "../../utils/errorResponse";
 import { respond } from "../../utils/respond";
 import { getErrorMessage } from './shared';
