--- core/src/services/AuthService.ts
@@ -1,1 +1,0 @@
-export * from "../domains/identity/application/auth/AuthService";
