--- core/src/services/UserProfileService.ts
@@ -1,1 +1,0 @@
-export * from "../domains/identity/application/users/UserProfileService";
