--- core/src/services/UserService.ts
@@ -1,1 +1,0 @@
-export * from "../domains/identity/application/users/UserService";
