--- core/src/services/UserStatusService.ts
@@ -1,1 +1,0 @@
-export * from "../domains/identity/application/users/UserStatusService";
